feat(caPatchService): add informative error message as default parameter for listener
The default parameter for the listener function patched in is now a function that throws an
error telling the type of attribute being manipulated in the controller and the line in the
controllower where the manipulation occurs. The parameter can be overwritten if
patchServices.listener is set to an alternate function.

Closes #8