test(): add test to try and diagnose async issue
An issue has developed with an error being thrown
after an asynchronous call completes. This test aimed
to show if some prototype remained unpatched. Here, the
listener is only called when expected, inside the timeout.
All other manipulations do not call the listener, illustrating
that the prototypes are unpatched after that point.

#22